refactor: 패키지 구조 변경
domain 패키지 생성
