+++ src/main/java/lotto/domain/RandomLottoTicketGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
